# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 3D Phone Store
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Phone {
  id: number;
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  /** @nullable */
  imageUrl2?: string | null;
  /** @nullable */
  imageUrl3?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  specs?: string | null;
  categoryId: number;
  featured?: boolean;
  inStock?: boolean;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  storage?: string | null;
}

export interface PhoneInput {
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  /** @nullable */
  imageUrl2?: string | null;
  /** @nullable */
  imageUrl3?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  specs?: string | null;
  categoryId: number;
  featured?: boolean;
  inStock?: boolean;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  storage?: string | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}

export interface CartItem {
  id: number;
  phoneId: number;
  quantity: number;
  phone: Phone;
}

export interface CartItemInput {
  phoneId: number;
  quantity: number;
}

export interface StoreStats {
  totalProducts: number;
  totalBrands: number;
  minPrice: number;
  maxPrice: number;
  inStockCount: number;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  role: string;
}

export interface OrderItem {
  id: number;
  phoneId: number;
  quantity: number;
  price: number;
  phone?: Phone;
}

export interface Order {
  id: number;
  userId: number;
  status: string;
  total: number;
  createdAt: string;
  /** @nullable */
  username?: string | null;
  items: OrderItem[];
}

export interface OrderStatusInput {
  status: string;
}

export type ListPhonesParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
featured?: boolean | null;
};