# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for 3D Phone Store
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: phones
    description: Phone products
  - name: categories
    description: Phone categories
  - name: cart
    description: Shopping cart
  - name: auth
    description: Authentication
  - name: orders
    description: Orders
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /phones:
    get:
      operationId: listPhones
      tags: [phones]
      summary: List all phones
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of phones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phone"
    post:
      operationId: createPhone
      tags: [phones]
      summary: Create a phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phone"

  /phones/{id}:
    get:
      operationId: getPhone
      tags: [phones]
      summary: Get a phone by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Phone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phone"
        "404":
          description: Not found
    delete:
      operationId: deletePhone
      tags: [phones]
      summary: Delete a phone (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden

  /phones/featured:
    get:
      operationId: getFeaturedPhones
      tags: [phones]
      summary: Get featured phones for hero display
      responses:
        "200":
          description: Featured phones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phone"

  /phones/stats:
    get:
      operationId: getPhoneStats
      tags: [phones]
      summary: Get store stats
      responses:
        "200":
          description: Store stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"

  /cart/{id}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not logged in

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place order from cart
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/orders:
    get:
      operationId: listAllOrders
      tags: [admin]
      summary: List all orders (admin)
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/orders/{id}:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Phone:
      type: object
      required: [id, name, brand, price, imageUrl, categoryId]
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        imageUrl2:
          type: ["string", "null"]
        imageUrl3:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        specs:
          type: ["string", "null"]
        categoryId:
          type: integer
        featured:
          type: boolean
        inStock:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        color:
          type: ["string", "null"]
        storage:
          type: ["string", "null"]

    PhoneInput:
      type: object
      required: [name, brand, price, imageUrl, categoryId]
      properties:
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        imageUrl2:
          type: ["string", "null"]
        imageUrl3:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        specs:
          type: ["string", "null"]
        categoryId:
          type: integer
        featured:
          type: boolean
        inStock:
          type: boolean
        color:
          type: ["string", "null"]
        storage:
          type: ["string", "null"]

    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]

    CartItem:
      type: object
      required: [id, phoneId, quantity, phone]
      properties:
        id:
          type: integer
        phoneId:
          type: integer
        quantity:
          type: integer
        phone:
          $ref: "#/components/schemas/Phone"

    CartItemInput:
      type: object
      required: [phoneId, quantity]
      properties:
        phoneId:
          type: integer
        quantity:
          type: integer

    StoreStats:
      type: object
      required: [totalProducts, totalBrands, minPrice, maxPrice, inStockCount]
      properties:
        totalProducts:
          type: integer
        totalBrands:
          type: integer
        minPrice:
          type: number
        maxPrice:
          type: number
        inStockCount:
          type: integer

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthUser:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string

    Order:
      type: object
      required: [id, userId, status, total, createdAt, items]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
        total:
          type: number
        createdAt:
          type: string
        username:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderItem:
      type: object
      required: [id, phoneId, quantity, price]
      properties:
        id:
          type: integer
        phoneId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        phone:
          $ref: "#/components/schemas/Phone"

    OrderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string