# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 3D Phone Store
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  userId: number;
  status: string;
  total: number;
  createdAt: string;
  /** @nullable */
  username?: string | null;
  items: OrderItem[];
}