# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 3D Phone Store
 * OpenAPI spec version: 0.1.0
 */
import type { Phone } from './phone';

export interface OrderItem {
  id: number;
  phoneId: number;
  quantity: number;
  price: number;
  phone?: Phone;
}