# lib/api-zod/src/generated/types/orderStatusInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 3D Phone Store
 * OpenAPI spec version: 0.1.0
 */

export interface OrderStatusInput {
  status: string;
}